requirements.txt
numpy==1.25.2
Pillow==9.5.0
streamlit==1.25.0

